Move repo resolution to `mini_git::core::repository`
