+++ src/core/commands/mod.rs
@@ -7,124 +7,10 @@ pub mod ls_tree;
 pub mod rev_parse;
 pub mod show_ref;
 
-use std::collections::HashSet;
-use std::fs;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 
-use crate::core::objects::{self, tree, worktree};
+use crate::core::objects::worktree;
 use crate::core::GitRepository;
-use crate::utils::path;
-
-/// Represents the source of a file, either from a Git blob or the working tree.
-#[derive(Debug)]
-pub enum FileSource {
-    /// A file stored in a Git blob, with a specific path and SHA identifier.
-    Blob { path: String, sha: String },
-
-    /// A file located in the working tree with a specified path.
-    Worktree { path: String },
-}
-
-/// Holds the context of a Git repository, including the current working directory,
-/// repository path, and a reference to the Git repository.
-#[derive(Debug)]
-pub struct RepositoryContext {
-    /// The current working directory, resolved when `resolve_repository_context` is called.
-    pub cwd: PathBuf,
-
-    /// The absolute path to the root of the repository's worktree.
-    pub repo_path: PathBuf,
-
-    /// The `GitRepository` representing the current repository.
-    pub repo: GitRepository,
-}
-
-impl FileSource {
-    /// Retrieves the contents of the file source.
-    ///
-    /// # Arguments
-    ///
-    /// * `repo` - Reference to the Git repository.
-    ///
-    /// # Returns
-    ///
-    /// A `Result` containing the file contents as a vector of bytes if successful,
-    /// or an error message if reading the contents failed.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if:
-    ///   - For `Blob` sources, the object is not a blob.
-    ///   - For `Worktree` sources, the file could not be read from the filesystem.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use mini_git::core::commands::{FileSource, RepositoryContext, resolve_repository_context};
-    ///
-    /// let RepositoryContext { repo, .. } = resolve_repository_context()?;
-    ///
-    /// let file_source = FileSource::Blob { path: "file.txt".to_string(), sha: "abc123".to_string() };
-    /// let contents = file_source.contents(&repo)?;
-    ///
-    /// # Ok::<(), String>(())
-    /// ```
-    pub fn contents(&self, repo: &GitRepository) -> Result<Vec<u8>, String> {
-        Ok(match self {
-            FileSource::Blob { sha, .. } => {
-                match objects::read_object(repo, sha)? {
-                    objects::GitObject::Blob(blob) => blob.data,
-                    x => {
-                        return Err(format!(
-                            "Expect object {sha} to be a blob, but was {}",
-                            String::from_utf8_lossy(x.format())
-                        ))
-                    }
-                }
-            }
-            FileSource::Worktree { path } => match fs::read(path) {
-                Ok(data) => data,
-                Err(e) => {
-                    return Err(format!(
-                        "Failed to read file {path}! Error: {e}"
-                    ))
-                }
-            },
-        })
-    }
-
-    /// Returns the path of the file, either from a Git blob or working tree.
-    ///
-    /// # Returns
-    ///
-    /// A `String` representing the path to the file.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::path::Path;
-    /// use mini_git::core::commands::FileSource;
-    ///
-    /// let file_source = FileSource::Worktree { path: "file.txt".to_string() };
-    /// assert_eq!(file_source.path(), "file.txt");
-    /// ```
-    #[must_use]
-    pub fn path(&self) -> String {
-        match self {
-            FileSource::Blob { path, .. } | FileSource::Worktree { path } => {
-                path.clone()
-            }
-        }
-    }
-}
-
-impl AsRef<Path> for FileSource {
-    fn as_ref(&self) -> &Path {
-        use FileSource::{Blob, Worktree};
-        let (Worktree { ref path } | Blob { ref path, .. }) = self;
-        Path::new(path.as_str())
-    }
-}
 
 /// Resolves files specified on the command line to paths relative to the repository root.
 ///
@@ -145,7 +31,8 @@ impl AsRef<Path> for FileSource {
 ///
 /// # Examples
 /// ```
-/// use mini_git::core::commands::{RepositoryContext, resolve_cla_files, resolve_repository_context};
+/// use mini_git::core::{RepositoryContext, resolve_repository_context};
+/// use mini_git::core::commands::resolve_cla_files;
 /// let RepositoryContext {cwd, repo, ..} = resolve_repository_context()?;
 ///
 /// let resolved_files = resolve_cla_files(&repo, &cwd, "src/main.rs,src/lib.rs")?;
@@ -203,90 +90,4 @@ pub fn resolve_cla_files(
     }
 
     Ok(resolved_files)
-}
-
-/// Retrieves files from a specified tree or the working directory if no tree is specified.
-///
-/// # Parameters
-/// - `repo`: A reference to the `GitRepository`.
-/// - `tree`: An optional reference to a tree identifier.
-///
-/// # Returns
-/// - `Ok(Vec<FileSource>)` containing files from the specified tree or working directory.
-/// - `Err(String)` if an error occurs while retrieving files.
-///
-/// # Errors
-/// - Returns an error if:
-///   - Files cannot be read from the specified tree.
-///   - The working directory cannot be accessed.
-fn get_files(
-    repo: &GitRepository,
-    tree: Option<&str>,
-) -> Result<Vec<FileSource>, String> {
-    Ok(match tree {
-        // Get contents from the specified tree
-        Some(treeish) => {
-            tree::get_tree_files(repo, treeish)?.into_iter().collect()
-        }
-
-        // Get contents from the working directory
-        None => worktree::get_worktree_files(repo, None)?
-            .into_iter()
-            .collect(),
-    })
-}
-
-/// Collects all files that need to be processed, based on user-specified files or default paths.
-///
-/// # Parameters
-/// - `files1`: A slice of `FileSource` representing files from the first tree.
-/// - `files2`: A slice of `FileSource` representing files from the second tree.
-/// - `specified_files`: A slice of `String` with specific files to process, if any.
-///
-/// # Returns
-/// A `Vec<String>` containing paths to all files that need processing.
-fn collect_files_to_process(
-    files1: &[FileSource],
-    files2: &[FileSource],
-    specified_files: &[String],
-) -> Vec<String> {
-    let mut all_files = HashSet::new();
-
-    if specified_files.is_empty() {
-        all_files.extend(files1.iter().map(FileSource::path));
-        all_files.extend(files2.iter().map(FileSource::path));
-    } else {
-        all_files.extend(specified_files.iter().cloned());
-    }
-
-    all_files.into_iter().collect()
-}
-
-/// Resolves the repository context, including the current working directory, repository path,
-/// and repository object.
-///
-/// # Returns
-/// - `Ok(RepositoryContext)` containing the current working directory, repository path, and Git repository object.
-/// - `Err(String)` if the repository context cannot be determined.
-///
-/// # Errors
-/// - Returns an error if:
-///   - The current working directory cannot be determined.
-///   - The repository path cannot be determined.
-///   - The Git repository object cannot be initialized.
-pub fn resolve_repository_context() -> Result<RepositoryContext, String> {
-    let cwd = std::env::current_dir().map_err(|_| {
-        "Could not determine current working directory".to_owned()
-    })?;
-
-    let repo_path = path::repo_find(&cwd)?
-        .canonicalize()
-        .map_err(|_| "Could not determine repository path".to_owned())?;
-    let repo = GitRepository::new(&repo_path)?;
-
-    Ok(RepositoryContext {
-        cwd,
-        repo_path,
-        repo,
-    })
 }
